Added library for seperate function definitions.
